Added comment
To show where the error is and why it is happening, this error message only shows on the localhost version and not the live copy for some reason. So to test use the localhost, then when it seems to have worked ( get the thankyou screen) then upload the files to the site and test.